Add new top menu to the app + transforme lanfing page from modal to a page
