+++ package.json
@@ -36,7 +36,7 @@
     "ajv": "8.8.2",
     "ajv-keywords": "5.1.0",
     "ang-jsoneditor": "1.10.5",
-    "arlas-wui-toolkit": "24.3.0-beta.4",
+    "arlas-wui-toolkit": "24.3.0-beta.5",
     "file-saver": "^2.0.2",
     "jsoneditor": "9.5.6",
     "moment": "^2.29.4",
