+++ src/app/components/landing-page/landing-page-dialog.component.scss
@@ -73,9 +73,7 @@ $primary-color: mat-color($primary-palette);
   }
 }
 
-.new-config-content {
-  overflow: unset;
-}
+
 
 .actions {
   .buttons {
