+++ src/app/components/landing-page/landing-page.component.html
@@ -1,1 +1,58 @@
 <arlas-tool-errormodal></arlas-tool-errormodal>
+<ngx-spinner [name]="'importconfig'" bdOpacity="0.2" bdColor="rgba(51,51,51,0.6)" size="medium" color="#fff"
+    type="ball-clip-rotate-pulse">
+    <p style="color: white"> {{'Importing dashboard...' | translate }} </p>
+</ngx-spinner>
+
+<h2 mat-dialog-title *ngIf="this.persistenceService.isAvailable" class="title">
+    <div>{{'Available dashboards' | translate}}</div>
+    <div *ngIf="this.isAuthenticated && this.orgs.length > 0" class="organisations">
+       <mat-form-field appearance="outline">
+        <mat-label>{{'organisation'}}</mat-label>
+        <mat-select (selectionChange)="changeOrg($event)" [value]="currentOrga" >
+            <mat-option *ngFor="let org of orgs" value="{{org.name}}">{{org.displayName}}</mat-option>
+        </mat-select>
+       </mat-form-field>
+    </div>
+</h2>
+<mat-dialog-content class="load-content">
+    <ng-container *ngIf="this.persistenceService.isAvailable">
+        <table mat-table [dataSource]="configurations">
+
+            <ng-container matColumnDef="id">
+                <th mat-header-cell *matHeaderCellDef>{{'Name' | translate}}</th>
+                <td mat-cell *matCellDef="let config">{{config.name}}</td>
+            </ng-container>
+
+            <ng-container matColumnDef="creation">
+                <th mat-header-cell *matHeaderCellDef>{{'Last update' | translate}}</th>
+                <td mat-cell *matCellDef="let config">
+                    <ng-container *ngIf="config.last_update_date != 0">{{config.last_update_date | date:'short'}}
+                    </ng-container>
+                </td>
+            </ng-container>
+
+            <ng-container matColumnDef="detail">
+                <th mat-header-cell *matHeaderCellDef>{{'Actions' | translate}}</th>
+                <td mat-cell *matCellDef="let config">
+                    <arlas-config-menu matTooltip="{{'Menu' | translate}}" [zone]="'config.json'"
+                        [actions]="config?.actions" (actionExecutedEmitter)="afterAction($event)">
+                    </arlas-config-menu>
+                </td>
+            </ng-container>
+
+            <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+            <tr mat-row *matRowDef="let config; columns: displayedColumns;"></tr>
+        </table>
+        <mat-paginator [pageIndex]="this.configPageNumber" [pageSize]="this.configPageSize"
+            [length]="configurationsLength" [showFirstLastButtons]=true (page)="pageChange($event)">
+        </mat-paginator>
+    </ng-container>
+</mat-dialog-content>
+
+<div class="buttons">
+    <button mat-stroked-button color="primary" class="new_dashboard" (click)="openChoice(InitialChoice.setup)">{{'New dashboard' |
+        translate}}</button>
+    <button mat-stroked-button color="primary" (click)="openChoice(InitialChoice.load)">{{'Import from file'
+        | translate}}</button>
+</div>
