+++ src/app/components/landing-page/landing-page.component.ts
@@ -16,40 +16,31 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-import { HttpClient } from '@angular/common/http';
-import { AfterViewInit, Component, Inject, OnDestroy, OnInit, Output } from '@angular/core';
-import { MAT_DIALOG_DATA, MatDialog, MatDialogRef } from '@angular/material/dialog';
+import { AfterViewInit, Component, OnDestroy, OnInit } from '@angular/core';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { PageEvent } from '@angular/material/paginator';
 import { MatSelectChange } from '@angular/material/select';
 import { ActivatedRoute, Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
-import { CollectionService } from '@services/collection-service/collection.service';
-import { DefaultValuesService } from '@services/default-values/default-values.service';
-import { MainFormManagerService } from '@services/main-form-manager/main-form-manager.service';
 import { Config } from '@services/main-form-manager/models-config';
 import { MapConfig } from '@services/main-form-manager/models-map-config';
 import { MainFormService } from '@services/main-form/main-form.service';
 import { MenuService } from '@services/menu/menu.service';
 import { StartingConfigFormBuilderService } from '@services/starting-config-form-builder/starting-config-form-builder.service';
 import { StartupService, ZONE_WUI_BUILDER } from '@services/startup/startup.service';
-import { DialogData } from '@shared-components/input-modal/input-modal.component';
 import { UserOrgData } from 'arlas-iam-api';
 import { DataWithLinks } from 'arlas-persistence-api';
 import {
-  ArlasAuthentificationService,
-  ArlasConfigService, ArlasConfigurationDescriptor, ArlasIamService, ArlasSettingsService, AuthentificationService, ConfigAction,
+  ArlasAuthentificationService, ArlasIamService, ArlasSettingsService, AuthentificationService, ConfigAction,
   ConfigActionEnum, ErrorService, PersistenceService, UserInfosComponent
 } from 'arlas-wui-toolkit';
 import { NGXLogger } from 'ngx-logger';
-import { NgxSpinnerService } from 'ngx-spinner';
-import { Subject, Subscription } from 'rxjs';
+import { Subscription } from 'rxjs';
 import { map } from 'rxjs/internal/operators/map';
+import { LandingPageDialogComponent } from './landing-page-dialog.component';
+import { LandingPageService } from '@services/landing-page/landing-page.service';
 
-enum InitialChoice {
-  none = 0,
-  setup = 1,
-  load = 2
-}
+
 
 export interface Configuration {
   name: string;
@@ -57,86 +48,68 @@ export interface Configuration {
   actions: Array<ConfigAction>;
 }
 
-@Component({
-  templateUrl: './landing-page-dialog.component.html',
-  styleUrls: ['./landing-page-dialog.component.scss']
-})
-export class LandingPageDialogComponent implements OnInit, OnDestroy {
-  @Output() public startEvent: Subject<boolean> = new Subject<boolean>();
+enum InitialChoice {
+  none = 0,
+  setup = 1,
+  load = 2
+}
 
-  public configChoice = InitialChoice.none;
-  public isServerReady = false;
-  public availablesCollections: string[];
-  public InitialChoice = InitialChoice;
+@Component({
+  selector: 'arlas-landing-page',
+  templateUrl: './landing-page.component.html',
+  styleUrls: ['./landing-page.component.scss']
+})
+export class LandingPageComponent implements OnInit, AfterViewInit, OnDestroy {
+
+  public dialogRef: MatDialogRef<LandingPageDialogComponent>;
+  public displayedColumns: string[] = ['id', 'creation', 'detail'];
+
+  private confId = '-1';
 
   public configurations: Configuration[] = [];
-  public displayedColumns: string[] = ['id', 'creation', 'detail'];
   public configurationsLength = 0;
   public configPageNumber = 0;
   public configPageSize = 5;
+  private errorAlreadyThrown = false;
+  public configChoice;
+  public availablesCollections: string[];
+  public InitialChoice = InitialChoice;
 
-  public showLoginButton: boolean;
-  public showLogOutButton: boolean;
   public authentMode = 'false';
 
   public isAuthenticated = false;
   public isAuthentActivated = false;
-  private errorAlreadyThrown = false;
-  public name: string;
-  public avatar: string;
   public orgs: UserOrgData[] = [];
   public currentOrga = '';
 
   private subscription: Subscription;
-  private urlSubscription: Subscription;
-  private urlCollectionsSubscription: Subscription;
-  private collectionsSubscription: Subscription;
   private refreshSubscription: Subscription;
-
   public constructor(
-    @Inject(MAT_DIALOG_DATA) public data: DialogData,
-    public dialogRef: MatDialogRef<LandingPageDialogComponent>,
-    public mainFormService: MainFormService,
-    private http: HttpClient,
-    private logger: NGXLogger,
-    private configService: ArlasConfigService,
     public startupService: StartupService,
-    private configDescritor: ArlasConfigurationDescriptor,
-    private startingConfigFormBuilder: StartingConfigFormBuilderService,
-    private defaultValuesService: DefaultValuesService,
-    private translate: TranslateService,
-    private mainFormManager: MainFormManagerService,
-    private collectionService: CollectionService,
     public persistenceService: PersistenceService,
+    public mainFormService: MainFormService,
+    private startingConfigFormBuilder: StartingConfigFormBuilderService,
+    private arlasAuthentService: ArlasAuthentificationService,
     private dialog: MatDialog,
     private authService: AuthentificationService,
     private errorService: ErrorService,
-    private arlasSettingsService: ArlasSettingsService,
-    private spinner: NgxSpinnerService,
-    private router: Router,
-    private menu: MenuService,
+    private settingsService: ArlasSettingsService,
     private arlasIamService: ArlasIamService,
-    private arlasAuthentService: ArlasAuthentificationService
-  ) {
+    private logger: NGXLogger,
+    private router: Router,
+    private translate: TranslateService,
+    private activatedRoute: ActivatedRoute,
+    private menu: MenuService,
+    private landingPageService: LandingPageService) {
 
-    this.isAuthentActivated = !!this.arlasAuthentService.authConfigValue && !!this.arlasAuthentService.authConfigValue.use_authent;
-    this.authentMode = this.arlasAuthentService.authConfigValue.auth_mode;
-    this.showLoginButton = this.isAuthentActivated && !this.isAuthenticated;
-    this.showLogOutButton = this.isAuthentActivated && this.isAuthenticated;
+    this.authentMode = !!this.settingsService.getAuthentSettings() ? this.settingsService.getAuthentSettings().auth_mode : undefined;
+    this.isAuthentActivated = !!this.settingsService.getAuthentSettings() && !!this.settingsService.getAuthentSettings().use_authent;
+    if (this.isAuthentActivated && !this.authentMode) {
+      this.authentMode = 'openid';
+    }
   }
 
   public ngOnInit(): void {
-    if (this.data.message !== '-1') {
-      this.loadConfig(this.data.message);
-    }
-    if (!!this.data.configChoice) {
-      if (this.data.configChoice === '2') {
-        this.configChoice = InitialChoice.load;
-      } else {
-        this.configChoice = InitialChoice.none;
-      }
-    }
-
     // Reset and clean the content of all forms
     this.mainFormService.resetMainForm();
 
@@ -149,33 +122,30 @@ export class LandingPageDialogComponent implements OnInit, OnDestroy {
     this.mainFormService.startingConfig.init(
       this.startingConfigFormBuilder.build()
     );
-
+    if (this.activatedRoute.snapshot.paramMap.has('id')) {
+      this.confId = this.activatedRoute.snapshot.paramMap.get('id');
+      this.loadConfig(this.confId);
+    } else {
+      if (
+        this.persistenceService.isAvailable
+        && (!this.arlasAuthentService.authConfigValue || !this.arlasAuthentService.authConfigValue.use_authent)
+      ) {
+        this.getConfigList();
+      }
+    }
     if (this.authentMode === 'openid') {
       const claims = this.authService.identityClaims as any;
       this.subscription = this.authService.canActivateProtectedRoutes.subscribe(isAuthenticated => {
-        // show login button when authentication is enabled in settings.yaml file && the app is not authenticated
-        this.showLoginButton = this.isAuthentActivated && !isAuthenticated;
-        this.showLogOutButton = this.isAuthentActivated && isAuthenticated;
         this.isAuthenticated = isAuthenticated;
         if (this.persistenceService.isAvailable) {
           this.getConfigList();
         }
-        if (isAuthenticated) {
-          this.name = claims.nickname;
-          this.avatar = claims.picture;
-        } else {
-          this.name = '';
-          this.avatar = '';
-        }
       });
-    }
-    if (this.authentMode === 'iam') {
+    } else if (this.authentMode === 'iam') {
       this.refreshSubscription = this.arlasIamService.tokenRefreshed$.subscribe({
         next: (loginData) => {
           if (!!loginData) {
             this.isAuthenticated = true;
-            this.name = loginData?.user.email;
-            this.avatar = '';
             this.orgs = loginData.user.organisations.map(org => {
               org.displayName = org.name === loginData.user.id ? loginData.user.email.split('@')[0] : org.name;
               return org;
@@ -183,14 +153,10 @@ export class LandingPageDialogComponent implements OnInit, OnDestroy {
             this.currentOrga = this.arlasIamService.getOrganisation();
           } else {
             this.isAuthenticated = false;
-            this.name = '';
-            this.avatar = '';
           }
           if (this.persistenceService.isAvailable) {
             this.getConfigList();
           }
-          this.showLoginButton = this.isAuthentActivated && !this.isAuthenticated;
-          this.showLogOutButton = this.isAuthentActivated && this.isAuthenticated;
         },
         error: () => {
           this.isAuthenticated = false;
@@ -199,184 +165,28 @@ export class LandingPageDialogComponent implements OnInit, OnDestroy {
     }
     // if persistence is configured and anonymous mode is enabled, we fetch the configuration accessible as anonymous
     // if ARLAS-persistence doesn't allow anonymous access, a suitable error is displayed in a modal
-    if (
-      this.persistenceService.isAvailable
-      && (!this.arlasAuthentService.authConfigValue || !this.arlasAuthentService.authConfigValue.use_authent)
-    ) {
-      this.getConfigList();
-    }
+
+    this.subscription = this.landingPageService.startEvent$.subscribe((b) => {
+      this.menu.updatePagesStatus(true);
+      this.router.navigate(['map-config'], { queryParamsHandling: 'preserve' });
+      this.logger.info(this.translate.instant('Ready to access server'));
+    });
   }
 
-  public ngOnDestroy() {
-    if (this.subscription) {
-      this.subscription.unsubscribe();
-    }
-    if (this.urlSubscription) {
-      this.urlSubscription.unsubscribe();
-    }
-    if (this.urlCollectionsSubscription) {
-      this.urlCollectionsSubscription.unsubscribe();
-    }
-    if (this.collectionsSubscription) {
-      this.collectionsSubscription.unsubscribe();
-    }
-    if (this.refreshSubscription) {
-      this.refreshSubscription.unsubscribe();
-    }
-  }
-
-  public logout() {
-    if (this.authentMode === 'openid') {
-      this.authService.logout();
-    }
-    if (this.authentMode === 'iam') {
-      this.arlasIamService.logout(['/']);
-    }
-  }
-
-  public cancel(): void {
-    this.dialogRef.close();
-  }
-  public onKeyUp(event: KeyboardEvent) {
-    // On press enter
-    if (event.key === 'Enter') {
-      this.checkUrl();
-    }
-  }
-
-  public checkUrl() {
-    this.spinner.show('connectServer');
-    const url = this.mainFormService.startingConfig.getFg().get('serverUrl').value;
-    this.urlSubscription = this.http.get(url + '/swagger.json').subscribe(
-      () => {
-        this.getServerCollections(url).then(
-          () => {
-            this.urlCollectionsSubscription = this.configDescritor.getAllCollections().subscribe(
-              collections => {
-                this.availablesCollections = collections;
-                this.collectionService.setCollections(collections);
-                this.collectionService.getCollectionsReferenceDescription().subscribe(cdrs => this.collectionService.setCollectionsRef(cdrs));
-              },
-              error => {
-                this.logger.error(this.translate.instant('Unable to access the server. Please, verify the url.'));
-                this.spinner.hide('connectServer');
-              },
-              () => this.spinner.hide('connectServer')
-            );
-            this.isServerReady = true;
-          });
-      },
-      () => {
-        this.logger.error(this.translate.instant('Unable to access the server. Please, verify the url.'));
-        this.isServerReady = false;
-        this.spinner.hide('connectServer');
-      }
-    );
-  }
-
-  public saveConfig() {
-    const collection = this.dialogRef.componentInstance.mainFormService.startingConfig.getFg().customControls.collection.value;
-    this.startupService.setDefaultCollection(collection);
-    this.mainFormManager.initMainModulesForms(true);
-    setTimeout(() => this.startEvent.next(null), 100);
-  }
-
-  public initWithConfig(configJson: Config, configMapJson: MapConfig, configId?: string, configName?: string, isRetry?: boolean) {
-    this.spinner.show('importconfig');
-    this.getServerCollections(configJson.arlas.server.url).then(() => {
-      this.collectionsSubscription = this.configDescritor.getAllCollections().subscribe({
-        next: collections => {
-          this.initCollections(collections, configJson, configMapJson, configId, configName);
-        },
-        error: () => {
-          this.spinner.hide('importconfig');
-          if (!isRetry) {
-            configJson.arlas.server.url = this.defaultValuesService.getValue('global.serverUrl');
-            this.initWithConfig(configJson, configMapJson, configId, configName, true);
-          } else {
-            this.logger.error(
-              this.translate.instant('Server "')
-              + configJson.arlas.server.url
-              + this.translate.instant('" unreachable'));
-          }
-        }
+  public openChoice(configChoice?) {
+    this.configChoice = configChoice;
+    if (configChoice) {
+      this.dialogRef = this.dialog.open(LandingPageDialogComponent, {
+        disableClose: true, data:
+          { message: this.confId, configChoice }
       });
-    });
+    }
   }
 
-  public initCollections(collections: string[], configJson: Config, configMapJson: MapConfig, configId?: string, configName?: string) {
-    this.collectionService.setCollections(collections);
-    this.collectionService.getCollectionsReferenceDescription().subscribe(cdrs => {
-      this.collectionService.setCollectionsRef(cdrs);
-      const collection = (collections.find(c => c === configJson.arlas.server.collection.name));
-      if (!collection) {
-        this.spinner.hide('importconfig');
-        this.logger.error(
-          this.translate.instant('Collection ')
-          + configJson.arlas.server.collection.name
-          + this.translate.instant(' unknown. Available collections: ')
-          + collections.join(', '));
-
-      } else {
-        // tslint:disable-next-line:no-string-literal
-        if (this.arlasSettingsService.settings['use_time_filter']) {
-          this.startupService.getTimeFilter(collection, configJson.arlas.server.url, this.collectionService, this.arlasSettingsService)
-            .subscribe(timeFilter => {
-              this.startupService.applyArlasInterceptor(collection, timeFilter);
-              setTimeout(() => {
-                this.mainFormManager.doImport(configJson, configMapJson);
-                this.startEvent.next(null);
-                this.spinner.hide('importconfig');
-              }, 100);
-            });
-          /** hack in order to trigger the spinner 'importconfig'
-           * otherwise, we will think the application is not loading
-           */
-        } else {
-          setTimeout(() => {
-            this.mainFormManager.doImport(configJson, configMapJson);
-            this.startEvent.next(null);
-            this.spinner.hide('importconfig');
-          }, 100);
-        }
-
-      }
-      if (!!configId && !!configName) {
-        this.mainFormService.configChange.next({ id: configId, name: configName });
-      }
-    });
-  }
-
-  public upload(configFile: File, mapConfigFile: File) {
-
-    const readAndParsePromise = (file: File) => new Promise((resolve, reject) => {
-      const fileReader = new FileReader();
-      fileReader.onload = (e) => {
-        try {
-          resolve(JSON.parse(fileReader.result.toString()));
-        } catch (excp) {
-          reject(excp);
-        }
-      };
-      fileReader.onerror = (err) => reject(err);
-      fileReader.readAsText(file);
-    });
-
-    Promise.all([
-      readAndParsePromise(configFile), readAndParsePromise(mapConfigFile)
-    ]).then(values => {
-      const configJson = values[0] as Config;
-      const configMapJson = values[1] as MapConfig;
-
-      this.initWithConfig(configJson, configMapJson);
-
-    }).catch(err => this.logger.error(this.translate.instant('Could not load config files ') + err));
-  }
-
-  public navigateToCollection(): void {
-    this.dialogRef.close();
-    this.menu.updatePagesStatus(false);
-    this.router.navigate(['/collection']);
+  public ngAfterViewInit() {
+    // the set timeout is a hack to display dialog after 'callback' redirection. Otherwise the dialog is not opened in that case.
+    // https://stackoverflow.com/questions/779379/why-is-settimeoutfn-0-sometimes-useful
+    setTimeout(() => this.openChoice(), 0);
   }
 
   public afterAction(event: ConfigAction) {
@@ -456,7 +266,7 @@ export class LandingPageDialogComponent implements OnInit, OnDestroy {
           this.configurations = result[1] as Configuration[];
           this.configurations.forEach(c => {
             c.actions.filter(a => a.type === ConfigActionEnum.VIEW)
-              .map(a => a.url = this.arlasSettingsService.getArlasWuiUrl());
+              .map(a => a.url = this.settingsService.getArlasWuiUrl());
           });
         },
         error: (msg) => {
@@ -502,32 +312,12 @@ export class LandingPageDialogComponent implements OnInit, OnDestroy {
     }
   }
 
-  private getServerCollections(serverUrl: string) {
-    // Update config with new server url
-    // Usefull to use toolkit method
-    const currentConf = this.configService.getConfig();
-    const newConf = Object.assign(currentConf, { arlas: { server: { url: serverUrl } } });
-    this.configService.setConfig(newConf);
-    // Update collaborative search Service with the new url
-    return this.startupService.setCollaborativeService(newConf);
-  }
-
-  public login() {
-    if (this.authentMode === 'openid') {
-      this.authService.login();
-    }
-    if (this.authentMode === 'iam') {
-      this.dialogRef.close();
-      this.router.navigate(['login']);
-    }
-  }
-
   public loadConfig(id: string) {
     this.persistenceService.get(id).subscribe(data => {
       const configJson = JSON.parse(data.doc_value) as Config;
       if (configJson.arlas !== undefined) {
         const configMapJson = configJson.arlas.web.components.mapgl.input.mapLayers as MapConfig;
-        this.initWithConfig(configJson, configMapJson, id, data.doc_key);
+        this.landingPageService.initWithConfig(configJson, configMapJson, id, data.doc_key);
       } else {
         this.configChoice = InitialChoice.setup;
       }
@@ -544,60 +334,14 @@ export class LandingPageDialogComponent implements OnInit, OnDestroy {
     this.startupService.changeOrgHeader(event.value, this.arlasIamService.getAccessToken());
     this.getConfigList();
   }
-}
-
-@Component({
-  selector: 'arlas-landing-page',
-  templateUrl: './landing-page.component.html',
-  styleUrls: ['./landing-page.component.scss']
-})
-export class LandingPageComponent implements OnInit, AfterViewInit, OnDestroy {
-
-  @Output() public startEvent: Subject<boolean> = new Subject<boolean>();
-  public dialogRef: MatDialogRef<LandingPageDialogComponent>;
-
-  private confId = '-1';
-  private configChoice;
-
-  public constructor(
-    private dialog: MatDialog,
-    private logger: NGXLogger,
-    private router: Router,
-    private translate: TranslateService,
-    private activatedRoute: ActivatedRoute,
-    private menu: MenuService) {
-  }
-
-  public ngOnInit(): void {
-    if (this.activatedRoute.snapshot.paramMap.has('id')) {
-      this.confId = this.activatedRoute.snapshot.paramMap.get('id');
-    } else if (!!this.activatedRoute.snapshot.routeConfig && this.activatedRoute.snapshot.routeConfig.path === 'import') {
-      this.configChoice = '2';
-    }
-  }
-
-  public openChoice() {
-    this.dialogRef = this.dialog.open(LandingPageDialogComponent, {
-      disableClose: true, data:
-        { message: this.confId, configChoice: this.configChoice }
-    });
-    this.dialogRef.componentInstance.startEvent.subscribe(mode => this.startEvent.next(mode));
-  }
-
-  public ngAfterViewInit() {
-    // the set timeout is a hack to display dialog after 'callback' redirection. Otherwise the dialog is not opened in that case.
-    // https://stackoverflow.com/questions/779379/why-is-settimeoutfn-0-sometimes-useful
-    setTimeout(() => this.openChoice(), 0);
-    this.startEvent.subscribe(state => {
-      this.dialogRef.close();
-      this.menu.updatePagesStatus(true);
-      this.router.navigate(['map-config'], { queryParamsHandling: 'preserve' });
-      this.logger.info(this.translate.instant('Ready to access server'));
-    });
-  }
 
   public ngOnDestroy() {
-    this.startEvent.unsubscribe();
+    if (!!this.subscription) {
+      this.subscription.unsubscribe();
+    }
+    if (this.refreshSubscription) {
+      this.refreshSubscription.unsubscribe();
+    }
   }
 }
