+++ src/app/modules/map-config/map-config.component.scss
@@ -24,5 +24,4 @@ under the License.
 }
 
 .map-container{
-  padding-top: 10px;
 }
