+++ src/app/modules/timeline-config/components/global-timeline/global-timeline.component.scss
@@ -24,5 +24,4 @@ under the License.
 }
 
 .toggle-detailed-timeline {
-  padding-top: 20px;
 }
