+++ src/app/services/starting-config-form-builder/starting-config-form-builder.service.ts
@@ -50,6 +50,10 @@ export class StartingConfigFormGroup extends FormGroup {
     },
   };
 
+  public reset() {
+    this.customControls.collection.setValue('');
+  }
+
 }
 
 @Injectable({
