+++ src/variables.scss
@@ -40,3 +40,8 @@ $arlasColor: #f63366;
    padding-bottom: 3px;
    
 }
+$top-menu-height: 40px;
+$left-menu-width: 48px;
+$menu-border: solid 1px rgba(0, 0, 0, 0.12);
+$default-font-size: 20px;
+$sm-margin: 5px;
